refactor(proxy): isolate output turn policy
